Correct the measure of the delay of an erase-block erasure
When a call to ems_write() triggers the erasure of an erase-block, it returns a
long time before the end of the erasure (which takes about 1 sec). So, the
next call to ems_write() is delayed by about 1 sec. As a result, the erasure
time is underestimated and the time to write the next block is overestimted.

Solution: measure the time of an erasure at the second ems_write(), not the
first.